requirements.txt
numpy
pandas
yfinance
statsmodels
matplotlib
pyyaml
pytest
